form, year
mainly, this gets rid of the old PHP form in favor of a formspree form that can be run on a static page. i used [this](http://www.austinzentz.com/2015/06/24/contact-forms-in-static-sites.html) to validate and also send the form without directing users to another page. downside of this is that it doesn’t use the reCAPTCHA built into form spree. so if this form produces a lot of junk from webscrapers, you might have to turn off the AJAX submithandler and just have it send users to another page.

i also made it so the year in the footer updates browser-side in JS. we are PHP-free now!